Split stuff out into files in preparation for potential future better packaging?
