+++ certapps/certapps.go
@@ -21,320 +21,13 @@ import (
 	"appengine/mail"
 	"appengine/urlfetch"
 	"appengine/user"
-
-	"github.com/mjibson/appstats"
 )
-
-type Greeting struct {
-	Author  string
-	Content string
-	Date    time.Time
-}
-
-type AuthContext struct {
-	LogInOutLink string
-	LogInOutText string
-	LoggedIn     bool
-}
-
-type MainAppContext struct {
-	Member *Member
-	Team   *Team
-	//User   *user.User
-
-	AuthContext
-}
-
-type JSONContext struct {
-	Data string
-}
-
-type ErrorContext struct {
-	Message string
-	Error   bool
-}
-
-type RemindersToSend struct {
-	All                 bool
-	RespondedYesOrMaybe bool
-	Unsent              bool
-	NoResponse          bool
-}
-
-type Location struct {
-	Line1 string
-	Line2 string
-	City  string
-	State string
-	Zip   string
-
-	Latitude        float64
-	Longitude       float64
-	PublicLatitude  float64
-	PublicLongitude float64
-}
-
-type Audit struct {
-	Created    time.Time
-	CreatedBy  *datastore.Key
-	Modified   time.Time
-	ModifiedBy *datastore.Key
-
-	Key   *datastore.Key `datastore:"-"`
-	KeyID int64          `datastore:"-"`
-
-	Deleted bool
-}
-
-type Team struct {
-	Name string
-	Location
-
-	MembersEmail  string
-	OfficersEmail string
-
-	GoogleAPIKey string
-
-	Audit
-}
-
-type TeamMember struct {
-	TeamKey   *datastore.Key
-	MemberKey *datastore.Key
-
-	Audit
-}
-
-type Person struct {
-	FirstName string
-	LastName  string
-	Cell      string
-	HomePhone string
-
-	Email  string
-	Email2 string
-
-	Location
-	UserID string
-
-	Enabled bool
-
-	LastLogin time.Time
-
-	Audit
-}
-
-type Member struct {
-	ShowCell  bool
-	ShowEmail bool
-	OKToText  bool
-
-	RadioID string
-
-	HomeAddress *datastore.Key
-
-	//TODO: this would need to move to TeamMember when multiple teams was supported
-	CanLookup bool `datastore:"-"`
-	Town      bool
-	OEM       bool
-	Officer   bool
-	Active    bool
-
-	Person
-}
-
-type MemberLogin struct {
-	MemberKey *datastore.Key
-	Login     time.Time
-	IP        string
-}
-
-type Event struct {
-	TeamKey *datastore.Key
-
-	Summary     string
-	Description string
-
-	Deployment bool
-	Exercise   bool
-	Meeting    bool
-	Training   bool
-
-	MeetTime    time.Time
-	EventStart  time.Time
-	EventFinish time.Time
-
-	EventLocation   Location
-	ParkingLocation Location
-
-	Link1 string //Deployment: IAP, Meeting: Agenda
-	Link2 string //Deployment: Roster
-
-	Responses []*MemberEvent `datastore:"-"`
-
-	Audit
-}
-
-type MemberEvent struct {
-	MemberKey *datastore.Key
-	EventKey  *datastore.Key
-
-	Attending bool
-	Sure      bool
-
-	FirstReminder  time.Time
-	FirstResponded time.Time
-	FirstViewed    time.Time
-	LastReminder   time.Time
-	LastResponded  time.Time
-	LastViewed     time.Time
-	Arrive         time.Time
-	Depart         time.Time
-
-	Audit
-}
-
-type Reminder struct {
-	MemberEvent *datastore.Key
-
-	Subject string
-	Body    string
-
-	Sent time.Time
-
-	Audit
-}
-
-type Certification struct {
-	Name        string `json:"name"`
-	MonthsValid int64  `json:"monthsValid"`
-
-	Audit
-}
-
-type CertificationClass struct {
-	CertificationKey *datastore.Key
-	TeamKey          *datastore.Key
-
-	Name      string    `json:"name"`
-	Scheduled time.Time `json:"scheduled"`
-	Cancelled bool      `json:"cancelled"`
-
-	Location
-
-	Attendees []*datastore.Key `json:"-"`
-
-	MembersAttending []Member `json:"attendees", datastore:"-"`
-
-	Audit
-}
-
-type TrainingTopic struct {
-	CertificationKey *datastore.Key
-
-	Name string `json:"name"`
-
-	EffectiveDate time.Time `json:"effectiveDate"`
-	SunsetDate    time.Time `json:"sunsetDate"`
-
-	Audit
-}
-
-type MemberClassContent struct {
-	TrainingTopicKey *datastore.Key
-	MemberKey        *datastore.Key
-
-	Attended bool
-
-	Audit
-}
-
-type MemberCertification struct {
-	CertificationKey *datastore.Key
-	MemberKey        *datastore.Key
-
-	Effective time.Time
-
-	Audit
-}
-
-type TeamCertification struct {
-	TeamKey          *datastore.Key
-	CertificationKey *datastore.Key
-	Abbr             string
-	Hidden           bool
-
-	Audit
-}
-
-type CertificationAndTopics struct {
-	Certification *Certification
-	Topics        []*TrainingTopic
-}
-
-// API types
-
-type GoogleLocationResults struct {
-	Results []GoogleLocation
-	Status  string
-}
-
-type GoogleLocation struct {
-	Geometry          GoogleLocationGeometry
-	Formatted_address string
-}
-
-type GoogleLocationGeometry struct {
-	Location LatLong
-}
-
-type LatLong struct {
-	Lat float64
-	Lng float64
-}
-
-// end types
 
 var jsonTemplate = textT.Must(textT.New("json").Parse("{\"data\": {{.Data}} }"))
 var jsonErrTemplate = textT.Must(textT.New("jsonErr").Parse("{\"error\": true, {{.Data}} }"))
 var zeroDate = time.Date(1, 1, 1, 0, 0, 0, 0, time.UTC)
 
 var delayEmailReminders2 = delay.Func("sendEmailReminders2", sendEmailReminders)
-
-func init() {
-	http.Handle("/api/trainingTopic/save", appstats.NewHandler(apiTrainingTopicSave))
-	http.Handle("/api/certificationClass", appstats.NewHandler(apiCertificationClassGet))
-	http.Handle("/api/certificationClass/all", appstats.NewHandler(apiCertificationClassGetAll))
-	http.Handle("/api/certificationClass/save", appstats.NewHandler(apiCertificationClassSave))
-	http.Handle("/api/certificationClass/attendee/add", appstats.NewHandler(apiCertificationClassAttendeeAdd))
-
-	http.Handle("/api/member/search", appstats.NewHandler(apiMemberSearch))
-
-	http.Handle("/audit", appstats.NewHandler(audit))
-	http.Handle("/certification", appstats.NewHandler(certificationGet))
-	http.Handle("/certifications/all", appstats.NewHandler(certificationsGetAll))
-	http.Handle("/certification/save", appstats.NewHandler(certificationSave))
-	http.Handle("/eventA", appstats.NewHandler(eventA))
-	http.Handle("/event", appstats.NewHandler(event))
-	http.Handle("/event/reminders/send", appstats.NewHandler(remindersSend))
-	http.Handle("/event/save", appstats.NewHandler(eventSave))
-	http.Handle("/events", appstats.NewHandler(events))
-	http.Handle("/member", appstats.NewHandler(memberData))
-	http.Handle("/member/save", appstats.NewHandler(memberSave))
-	http.Handle("/response", appstats.NewHandler(response))
-	http.Handle("/response/save", appstats.NewHandler(responseSave))
-	http.Handle("/responses/create", appstats.NewHandler(responsesCreate))
-	http.Handle("/team", appstats.NewHandler(teamData))
-	http.Handle("/team/roster", appstats.NewHandler(teamRoster))
-	http.Handle("/team/roster/import", appstats.NewHandler(membersImport))
-	http.Handle("/fix/teamMembers/without/Memeber", appstats.NewHandler(fixData))
-	http.Handle("/fix/members/geocode", appstats.NewHandler(resaveMembers))
-	http.Handle("/address", appstats.NewHandler(locationLookupHandler))
-	http.Handle("/setup", appstats.NewHandler(initialSetup))
-	http.Handle("/", appstats.NewHandler(root))
-
-	rand.Seed(time.Now().UnixNano())
-}
 
 func initialSetup(c appengine.Context, w http.ResponseWriter, r *http.Request) {
 	var context interface{}
@@ -1277,23 +970,6 @@ func (m *Member) addToTeam(teamKey *datastore.Key, c appengine.Context) error {
 	_, tmErr := datastore.Put(c, tm.Key, tm)
 
 	return tmErr
-}
-
-func (a *Audit) setAudits(m *Member) {
-	a.ModifiedBy = m.Key
-	a.Modified = time.Now()
-
-	if a.CreatedBy == nil {
-		a.Created = a.Modified
-		a.CreatedBy = a.ModifiedBy
-	}
-}
-
-func (a *Audit) setKey(key *datastore.Key) {
-	a.Key = key
-	if key != nil {
-		a.KeyID = key.IntID()
-	}
 }
 
 func event(c appengine.Context, w http.ResponseWriter, r *http.Request) {
@@ -2795,129 +2471,6 @@ func (t *CertificationClass) lookup(id int64, member *Member, c appengine.Contex
 	return err
 }
 
-func apiCertificationClassSave(c appengine.Context, w http.ResponseWriter, r *http.Request) {
-	u := user.Current(c)
-	var context interface{}
-	var mem *Member
-	var postData struct {
-		CClass *CertificationClass
-		Team   *Team
-	}
-
-	c.Infof("certificationClassSave")
-
-	// err := parseJSON(postData, r, c)
-	// if noErrMsg(err, w, c, "Parsing JSON") {
-
-	decoder := json.NewDecoder(r.Body)
-	jsonDecodeErr := decoder.Decode(&postData)
-
-	if jsonDecodeErr == io.EOF {
-		c.Infof("EOF, should it be?")
-	} else if noErrMsg(jsonDecodeErr, nil, c, "Parsing json from body") {
-		c.Infof("JSON from request: %+v", postData)
-
-		mem, _ = getMemberFromUser(c, u, w, r)
-
-		teamKey := datastore.NewKey(c, "Team", "", postData.Team.KeyID, nil)
-		postData.CClass.TeamKey = teamKey
-
-		saveErr := postData.CClass.save(mem, c)
-
-		c.Debugf("CertificationClass after save: %+v", postData.CClass)
-
-		if noErrMsg(saveErr, w, c, "Saving Certification") {
-			context = struct {
-				CClass *CertificationClass
-			}{
-				postData.CClass,
-			}
-		}
-	} else {
-
-	}
-
-	returnJSONorErrorToResponse(context, c, w, r)
-}
-
-func apiCertificationClassGet(c appengine.Context, w http.ResponseWriter, r *http.Request) {
-	u := user.Current(c)
-	var context interface{}
-	var mem *Member
-
-	id := parseIntIDVal(r)
-
-	c.Infof("certificationClassGet %d", id)
-
-	mem, _ = getMemberFromUser(c, u, w, r)
-
-	cClass := new(CertificationClass)
-	err := cClass.lookup(id, mem, c)
-
-	if noErrMsg(err, w, c, "Certification lookup") {
-		context = struct {
-			CertificationClass *CertificationClass
-		}{
-			cClass,
-		}
-	}
-
-	returnJSONorErrorToResponse(context, c, w, r)
-}
-
-func apiCertificationClassGetAll(c appengine.Context, w http.ResponseWriter, r *http.Request) {
-	u := user.Current(c)
-	var context interface{}
-	var mem *Member
-
-	id := parseIntFormVal("teamID", r)
-
-	c.Infof("certificationClassGetAll, id: %d", id)
-
-	if id != 0 {
-		mem, _ = getMemberFromUser(c, u, w, r)
-
-		if mem != nil {
-			var results []*CertificationClass
-			var teamKey = datastore.NewKey(c, "Team", "", id, nil)
-
-			query := datastore.NewQuery("CertificationClass").Filter("TeamKey =", teamKey)
-
-			keys, err := query.GetAll(c, &results)
-
-			if noErrMsg(err, w, c, "Getting All CertificationClasses for Team") {
-				for idx, _ := range results {
-					e := results[idx]
-					key := keys[idx]
-
-					e.setKey(key)
-				}
-			}
-
-			context = struct {
-				CClasses []*CertificationClass
-			}{
-				results,
-			}
-		} else {
-			context = struct {
-				NotLoggedIn bool
-			}{
-				false,
-			}
-		}
-	} else {
-		context = struct {
-			Error   bool
-			Message string
-		}{
-			true,
-			"Must pass teamID",
-		}
-	}
-	returnJSONorErrorToResponse(context, c, w, r)
-}
-
 func apiMemberSearch(c appengine.Context, w http.ResponseWriter, r *http.Request) {
 	var context interface{}
 	queries := 0
@@ -3089,80 +2642,6 @@ func memberQuery(field string, value string, equality bool, c appengine.Context)
 	}
 
 	return members, err
-}
-
-func apiCertificationClassAttendeeAdd(c appengine.Context, w http.ResponseWriter, r *http.Request) {
-	u := user.Current(c)
-	var context interface{}
-
-	var postData struct {
-		CClass *CertificationClass
-		Member *Member
-	}
-
-	c.Infof("certificationClassAddAttendee")
-
-	mem, memberErr := getMemberFromUser(c, u, w, r)
-
-	if noErrMsg(memberErr, w, c, "getMemberFromUser") {
-		decoder := json.NewDecoder(r.Body)
-		jsonDecodeErr := decoder.Decode(&postData)
-
-		if jsonDecodeErr == io.EOF {
-			c.Infof("EOF, should it be?")
-		} else if noErrMsg(jsonDecodeErr, nil, c, "Parsing json from body") {
-			c.Infof("JSON from request: KeyIDs Member: %d, CClass: %d", postData.Member.KeyID, postData.CClass.KeyID)
-
-			classLookupErr := postData.CClass.lookup(postData.CClass.KeyID, mem, c)
-
-			if noErrMsg(classLookupErr, w, c, "Looking up Certification Class") {
-				postData.Member.Key = datastore.NewKey(c, "Member", "", postData.Member.KeyID, nil)
-
-				if postData.Member.Key.Incomplete() == false {
-
-					unique := true
-					for _, attendee := range postData.CClass.Attendees {
-						if attendee.IntID() == postData.Member.KeyID {
-							unique = false
-							break
-						}
-					}
-
-					if unique {
-						postData.CClass.Attendees = append(postData.CClass.Attendees, postData.Member.Key)
-
-						classSaveErr := postData.CClass.save(mem, c)
-
-						if noErrMsg(classSaveErr, w, c, "Saving CClass with Member") {
-							context = struct {
-								Added bool
-							}{
-								true,
-							}
-						}
-					}
-				} else {
-					c.Errorf("Member passed is not a valid Key ID, it is incomplete, ie 0")
-				}
-			} else {
-				context = struct {
-					NotLoggedIn bool
-				}{
-					false,
-				}
-			}
-		} else {
-			context = struct {
-				Error   bool
-				Message string
-			}{
-				true,
-				"Must pass teamID",
-			}
-		}
-	}
-
-	returnJSONorErrorToResponse(context, c, w, r)
 }
 
 func checkErr(err error, w http.ResponseWriter, c appengine.Context, msg string) bool {
